+++ packages/@coorpacademy-app-review/src/actions/index.ts
@@ -1,13 +1,10 @@
-import {validateSlide} from './data/slides';
 import {StoreToken} from './data/token';
 
-import {updateFinishedSlides} from './ui/finished-slides';
 import {navigateBack, navigateTo, NavigateBack, NavigateTo, StartApp} from './ui/navigation';
-import {updateReviewStatus} from './ui/review-status';
 import {ReceivedSlide} from './api/fetch-slide';
 import {ReceiveProgression} from './api/post-progression';
 import {ReceivedSkills} from './api/fetch-skills';
-import {updateSlidesOnNext} from './ui/slides';
+import {validateSlide} from './ui/slides';
 
 export type Action =
   | NavigateBack
@@ -22,7 +19,4 @@ export type Dispatchers = {
   navigateTo: typeof navigateTo;
   navigateBack: typeof navigateBack;
   validateSlide: typeof validateSlide;
-  updateSlidesOnNext: typeof updateSlidesOnNext;
-  updateReviewStatus: typeof updateReviewStatus;
-  updateFinishedSlides: typeof updateFinishedSlides;
 };
