+++ packages/@coorpacademy-app-review/src/actions/ui/slides.ts
@@ -1,14 +1,5 @@
-import {SlideIndexes} from '../../common';
-import {ProgressionFromAPI, SlideFromAPI} from '../../types/common';
+import {SlideFromAPI} from '../../types/common';
 import {AnswerUI} from '../../types/slides';
-
-// -----------------------------------------------------------------------------
-
-export const SET_FIRST_SLIDE = '@@slide/SET_FIRST_SLIDE' as const;
-export const UPDATE_SLIDES_ON_VALIDATION = '@@slide/UPDATE_SLIDES_ON_VALIDATION' as const;
-export const UPDATE_SLIDES_ON_NEXT = '@@slide/UPDATE_SLIDES_ON_NEXT' as const;
-
-// -----------------------------------------------------------------------------
 
 type SlideUIAnimations = 'unstack' | 'restack';
 
@@ -23,51 +14,22 @@ export type Slide = {
   nextContent?: Omit<Slide, 'endReview' | 'hidden' | 'position' | 'nextContent'>;
 };
 
-export type OnNextPayload = {
-  slideIndex: SlideIndexes;
-  newSlideContent: Slide;
-  numberOfFinishedSlides: number;
-};
-
-// -----------------------------------------------------------------------------
-
+export const SET_FIRST_SLIDE = '@@slide/SET_FIRST_SLIDE' as const;
 export type SetFirstSlide = {
   type: typeof SET_FIRST_SLIDE;
   payload: SlideFromAPI;
 };
-
-export type UpdateSlidesOnValidation = {
-  type: typeof UPDATE_SLIDES_ON_VALIDATION;
-  payload: {
-    slideFromAPI: SlideFromAPI;
-    progressionFromAPI: ProgressionFromAPI;
-    fetchedSlidesRefs: string[];
-    numberOfFinishedSlides: number;
-  };
-};
-
-export type UpdateSlidesOnNext = {
-  type: typeof UPDATE_SLIDES_ON_NEXT;
-  payload: OnNextPayload;
-};
-
-export type UISlidesAction = UpdateSlidesOnValidation | UpdateSlidesOnNext | SetFirstSlide;
-
-// -----------------------------------------------------------------------------
-
 export const setFirstSlide = (payload: SlideFromAPI): SetFirstSlide => ({
   type: SET_FIRST_SLIDE,
   payload
 });
 
-export const updateSlidesOnValidation = (
-  payload: UpdateSlidesOnValidation['payload']
-): UpdateSlidesOnValidation => ({
-  type: UPDATE_SLIDES_ON_VALIDATION,
-  payload
+export const VALIDATE_SLIDE = '@@slide/VALIDATE';
+export type ValidateSlide = {
+  type: typeof VALIDATE_SLIDE;
+};
+export const validateSlide = (): ValidateSlide => ({
+  type: VALIDATE_SLIDE
 });
 
-export const updateSlidesOnNext = (payload: OnNextPayload): UpdateSlidesOnNext => ({
-  type: UPDATE_SLIDES_ON_NEXT,
-  payload
-});
+export type UISlidesAction = SetFirstSlide;
